+++ src/StackExchange.Redis/RedisLiterals.cs
@@ -113,6 +113,7 @@ namespace StackExchange.Redis
             NODES = "NODES",
             NOSAVE = "NOSAVE",
             NOT = "NOT",
+            NOVALUES = "NOVALUES",
             NUMPAT = "NUMPAT",
             NUMSUB = "NUMSUB",
             NX = "NX",